Fix weird sprite issues by sorting sprites on diffirent levels inside layers Add yield to structures Split order seleciton into 4 partial classes to make it more readable
