+++ Assets/Creature/Creature.prefab
@@ -236,7 +236,7 @@ Transform:
   m_GameObject: {fileID: 3374013003364883131}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 2.5, y: 2.5, z: 2.5}
+  m_LocalScale: {x: 2, y: 2, z: 2}
   m_Children:
   - {fileID: 930703357700816054}
   - {fileID: 7185154708868162667}
