+++ Assets/Creature/Tasks/RemoveStructure.cs
@@ -22,7 +22,7 @@ public class RemoveStructure : TaskBase
     {
         if (Taskmaster.QueueComplete(SubTasks))
         {
-            foreach (var itemName in StructureController.Instance.StructureDataReference[Structure.Name].RequiredItemTypes)
+            foreach (var itemName in StructureController.Instance.StructureDataReference[Structure.Name].Require)
             {
                 MapGrid.Instance.GetCellAtCoordinate(Coordinates).AddContent(ItemController.Instance.GetItem(itemName).gameObject, true);
             }
