+++ Assets/Structures/Resources/Bush.json
@@ -5,5 +5,6 @@
   "TravelCost": "3",
   "StructureType": "Bush",
   "Scatter": "True",
-  "Buildable":  "False"
+  "Buildable":  "False",
+	"Yield":["1-3:Food"]
 }
