+++ Assets/Structures/Resources/StoneWall.json
@@ -2,7 +2,7 @@
 	"Name":"Stone Wall",
 	"SpriteName":"StoneWall",
 	"Tiled":"False",
-	"RequiredItemTypes":["Rock", "Rock","Rock", "Rock"],
+	"Require":["Rock", "Rock","Rock", "Rock"],
 	"TravelCost": "-1",
 	"StructureType": "Wall",	
 	"Scatter": "False",
