+++ Assets/Structures/Resources/Tree.json
@@ -5,5 +5,6 @@
 	"TravelCost": "-1",
 	"StructureType": "Tree",
 	"Scatter": "True",
-	"Buildable":  "False"
+	"Buildable":  "False",
+	"Yield":["10:Wood"]
 }
