+++ Assets/UI/OrderSelectionController.Designate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b26230672d9bb8243aed58ded71845e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
