+++ Assets/UI/OrderSelectionController.Stockpile.cs
@@ -1,0 +1,48 @@
+﻿using UnityEngine.UI;
+
+public partial class OrderSelectionController //.Stockpile
+{
+    public const string DefaultStockpileText = "Place Stockpile";
+    internal OrderButton StockpileButton;
+
+    private void StockpileClicked(string itemTypeName)
+    {
+        StockpileButton.Text = $"Place {itemTypeName} Stockpile";
+
+        CellClickOrder = cells =>
+        {
+            foreach (var cell in cells)
+            {
+                if (cell.Data.Stockpile == null && cell.TravelCost > 0)
+                {
+                    var stockpile = StockpileController.Instance.AddStockpile(itemTypeName);
+                    cell.AddContent(stockpile.gameObject);
+                }
+            }
+        };
+    }
+
+    private void StockpileTypeClicked()
+    {
+        GameController.Instance.SelectionPreference = SelectionPreference.CellOnly;
+        if (OrderTrayController.Instance.gameObject.activeInHierarchy)
+        {
+            DisableAndReset();
+            StockpileButton.Text = DefaultStockpileText;
+        }
+        else
+        {
+            EnableAndClear();
+
+            foreach (var item in ItemController.Instance.AllItemTypes.Values)
+            {
+                var button = Instantiate(OrderButtonPrefab, OrderTrayController.Instance.transform);
+                button.Button.onClick.AddListener(() => StockpileClicked(item.Data.ItemType));
+                button.name = $"Place {item.Data.ItemType} Stockpile";
+                button.Button.image.sprite = SpriteStore.Instance.GetSpriteByName(item.Data.SpriteName);
+                button.Button.image.type = Image.Type.Tiled;
+                button.Text = button.name;
+            }
+        }
+    }
+}
