+++ Assets/UI/OrderSelectionController.Stockpile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b1914c1bdf0b5a41b70c0495a5ae98d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
