+++ Assets/UI/OrderSelectionController.Structure.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41c83381958089e429b33ead869148a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
